+++ packages/signals/signals/src/core/middleware/network-signals-filter/network-signals-filter.ts
@@ -1,6 +1,30 @@
-import { NetworkSettingsConfig } from '.'
+import { Signal } from '@segment/analytics-signals-runtime'
 import { RegexLike } from '../../../types/settings'
-import { isSameDomain } from './helpers'
+import { SignalsMiddleware, SignalsMiddlewareContext } from '../../emitter'
+import { SignalsSettingsConfig } from '../../signals'
+import { isSameDomain } from '../../signal-generators/network-gen/helpers'
+
+export type NetworkSettingsConfigSettings = Pick<
+  SignalsSettingsConfig,
+  | 'networkSignalsAllowList'
+  | 'networkSignalsAllowSameDomain'
+  | 'networkSignalsDisallowList'
+>
+export class NetworkSettingsConfig {
+  networkSignalsAllowSameDomain: boolean
+  networkSignalsFilterList: NetworkSignalsFilterList
+  constructor({
+    networkSignalsAllowList,
+    networkSignalsDisallowList,
+    networkSignalsAllowSameDomain,
+  }: NetworkSettingsConfigSettings) {
+    this.networkSignalsAllowSameDomain = networkSignalsAllowSameDomain ?? true
+    this.networkSignalsFilterList = new NetworkSignalsFilterList(
+      networkSignalsAllowList,
+      networkSignalsDisallowList
+    )
+  }
+}
 
 class NetworkFilterListItem {
   regexes: RegexLike[]
@@ -65,11 +89,12 @@ export class NetworkSignalsFilterList {
     ])
   }
 
+  isDisallowed(url: string): boolean {
+    return this.disallowed.test(url) || this.disallowedDefaults.test(url)
+  }
+
   isAllowed(url: string): boolean {
-    const disallowed =
-      this.disallowed.test(url) || this.disallowedDefaults.test(url)
-    const allowed = this.allowed.test(url)
-    return allowed && !disallowed
+    return this.allowed.test(url)
   }
 
   getRegexes() {
@@ -93,10 +118,37 @@ export class NetworkSignalsFilter {
     const { networkSignalsFilterList, networkSignalsAllowSameDomain } =
       this.settings
 
-    const passesNetworkFilter = networkSignalsFilterList.isAllowed(url)
-    const allowedBecauseSameDomain =
-      networkSignalsAllowSameDomain && isSameDomain(url)
-    const allowed = passesNetworkFilter || allowedBecauseSameDomain
+    // anything that is disallowed takes precedence over the allow list.
+    if (networkSignalsFilterList.isDisallowed(url)) {
+      return false
+    }
+
+    const allowed =
+      // allowed because it's in the allow list
+      networkSignalsFilterList.isAllowed(url) ||
+      // allowed because it's the same domain
+      (networkSignalsAllowSameDomain && isSameDomain(url))
     return allowed
+  }
+}
+
+export class NetworkSignalsFilterMiddleware implements SignalsMiddleware {
+  private filter!: NetworkSignalsFilter
+
+  load(ctx: SignalsMiddlewareContext): void | Promise<void> {
+    this.filter = new NetworkSignalsFilter(ctx.unstableGlobalSettings.network)
+  }
+
+  private createMetadata = () => ({
+    filters: this.filter.settings.networkSignalsFilterList.getRegexes(),
+  })
+
+  process(signal: Signal): Signal | null {
+    if (signal.type === 'network') {
+      signal.metadata = this.createMetadata()
+      return this.filter.isAllowed(signal.data.url) ? signal : null
+    } else {
+      return signal
+    }
   }
 }